requirements.txt
huggingface_hub>=0.23.0
transformers>=4.40.0
accelerate>=0.30.0
safetensors
bitsandbytes>=0.43.0
peft
lycoris-lora==1.8.3
optimum-quanto==0.2.4
timm
einops
Pillow>=10.0
regex
pyyaml
tqdm
requests
gradio>=4.0.0
google-genai>=1.0.0
flask>=3.0.0
